refactor: update component imports to use shadcn-ui library for consistency and enhance layout with new LanguageSwitcher and ShareDialog components
