+++ src/components/error/error-boundary.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { Button } from '@/components/ui/button'
+import { Button } from '@/components/shadcn-ui/button'
 import { useEffect } from 'react'
 import { toast } from 'sonner'
