+++ src/components/features/language-switcher.tsx
@@ -1,16 +1,16 @@
 'use client'
 
+import { Globe } from 'lucide-react'
 import { useRouter } from 'next/navigation'
 import { useTransition } from 'react'
-import { Globe } from 'lucide-react'
 
+import { Button } from '@/components/shadcn-ui/button'
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuTrigger,
-} from '@/components/ui/dropdown-menu'
-import { Button } from '@/components/ui/button'
+} from '@/components/shadcn-ui/dropdown-menu'
 
 const LANGUAGES = [
   { label: '日本語', value: 'ja' },
