+++ src/components/layout/header/index.tsx
@@ -1,7 +1,7 @@
 'use client'
 
-import { LanguageSwitcher } from '@/components/common/language-switcher'
-import { Button } from '@/components/ui/button'
+import { LanguageSwitcher } from '@/components/features/language-switcher'
+import { Button } from '@/components/shadcn-ui/button'
 import { Moon, Sun } from 'lucide-react'
 import { useTheme } from 'next-themes'
 import Link from 'next/link'
