+++ src/components/shadcn-ui/carousel.tsx
@@ -6,7 +6,7 @@ import useEmblaCarousel, {
 import { ArrowLeft, ArrowRight } from 'lucide-react'
 import * as React from 'react'
 
-import { Button } from '@/components/ui/button'
+import { Button } from '@/components/shadcn-ui/button'
 import { cn } from '@/lib/utils'
 
 type CarouselApi = UseEmblaCarouselType[1]
